jobspot_backend/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-JWT-Extended==4.5.3
Flask-Migrate==4.0.5
Flask-CORS==4.0.0
# Remove psycopg2 temporarily for development
requests==2.31.0
beautifulsoup4==4.12.2
python-dotenv==1.0.0
alembic==1.12.0
gunicorn==21.2.0
selenium==4.12.0
webdriver-manager==4.0.0
APScheduler==3.10.4
python-dateutil==2.8.2

# Development dependencies
pytest==7.4.2
pytest-cov==4.1.0
black==23.7.0
flake8==6.1.0
isort==5.12.0